+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+### Added
+- AXI-Lite manager port to crossbar
+- `uart_wrapper` module to cleaner hierarchy
+
 ## [v0.1.5] - 2026-09-01
 
 ### Added